--- a/rocprim/include/rocprim/device/detail/config/device_merge_sort_block_merge.hpp
+++ b/rocprim/include/rocprim/device/detail/config/device_merge_sort_block_merge.hpp
@@ -1016,6 +1016,39 @@ struct default_merge_sort_block_merge_config<
     : merge_sort_block_merge_config<256, 1, (1 << 17) + 70000, 128, 128, 8>
 {};
 
+// Based on key_type = int64_t, value_type = custom_type<char, double>
+template<class key_type, class value_type>
+struct default_merge_sort_block_merge_config<
+    static_cast<unsigned int>(target_arch::unknown),
+    key_type,
+    value_type,
+    std::enable_if_t<(!bool(rocprim::is_floating_point<key_type>::value) && (sizeof(key_type) <= 8)
+                      && (sizeof(value_type) <= 16) && (sizeof(value_type) > 8))>>
+    : merge_sort_block_merge_config<256, 1, (1 << 17) + 70000, 128, 512, 2>
+{};
+
+// Based on key_type = int64_t, value_type = custom_type<char, double>
+template<class key_type, class value_type>
+struct default_merge_sort_block_merge_config<
+    static_cast<unsigned int>(target_arch::gfx908),
+    key_type,
+    value_type,
+    std::enable_if_t<(!bool(rocprim::is_floating_point<key_type>::value) && (sizeof(key_type) <= 8)
+                      && (sizeof(value_type) <= 16) && (sizeof(value_type) > 8))>>
+    : merge_sort_block_merge_config<256, 1, (1 << 17) + 70000, 128, 512, 2>
+{};
+
+// Based on key_type = int64_t, value_type = custom_type<char, double>
+template<class key_type, class value_type>
+struct default_merge_sort_block_merge_config<
+    static_cast<unsigned int>(target_arch::gfx90a),
+    key_type,
+    value_type,
+    std::enable_if_t<(!bool(rocprim::is_floating_point<key_type>::value) && (sizeof(key_type) <= 8)
+                      && (sizeof(value_type) <= 16) && (sizeof(value_type) > 8))>>
+    : merge_sort_block_merge_config<256, 1, (1 << 17) + 70000, 128, 512, 2>
+{};
+
 } // end namespace detail
 
 END_ROCPRIM_NAMESPACE