… .gitlab-ci.yml
… CHANGELOG.md
… benchmark/ConfigAutotuneSettings.cmake
… benchmark/benchmark_block_sort.cpp
… benchmark/benchmark_block_sort.parallel.hpp
… benchmark/benchmark_device_merge_sort_block_sort.cpp
… benchmark/benchmark_device_merge_sort_block_sort.parallel.hpp
… benchmark/benchmark_utils.hpp
… rocprim/include/rocprim/device/detail/config/device_merge_sort.hpp
… rocprim/include/rocprim/device/detail/config/device_radix_sort.hpp
… rocprim/include/rocprim/device/detail/device_merge_sort.hpp
… rocprim/include/rocprim/device/device_merge_sort.hpp
… rocprim/include/rocprim/device/device_merge_sort_config.hpp
… rocprim/include/rocprim/device/device_radix_sort.hpp
… rocprim/include/rocprim/device/specialization/device_radix_merge_sort.hpp
… rocprim/include/rocprim/intrinsics/warp_shuffle.hpp
… rocprim/include/rocprim/types/tuple.hpp
… rocprim/include/rocprim/warp/detail/warp_sort_stable.hpp
… scripts/autotune/create_optimization.py
… scripts/autotune/templates/mergesort_config_template
… scripts/autotune/templates/radixsort_config_template
… test/hip/test_hip_async_copy.cpp
… test/rocprim/CMakeLists.txt
… test/rocprim/test_block_histogram.kernels.hpp
… test/rocprim/test_block_load_store.hpp
… test/rocprim/test_block_radix_sort.kernels.hpp
… test/rocprim/test_block_reduce.kernels.hpp
… test/rocprim/test_block_scan.kernels.hpp
… test/rocprim/test_block_shuffle.hpp
… test/rocprim/test_block_sort.hpp