Development iteration + big update to block_sort_merge, improved performance of merge_sort (also specifically the ssbk case).

I recommend merging this in develop before splitting off a 5.6 branch from develop, the next develop_stream PR (after this one) will contain 6.0 specific changes.

Performance on MI100:
Before (current develop branch)
```cpp
device,merge_sort,int,empty_type,default_config/manual_time                                              69.2 ms         69.2 ms           10 bytes_per_second=18.0635G/s items_per_second=4.84888G/s
device,merge_sort,int64_t,empty_type,default_config/manual_time                                           127 ms          127 ms            6 bytes_per_second=19.7186G/s items_per_second=2.64659G/s
device,merge_sort,int8_t,empty_type,default_config/manual_time                                           44.1 ms         44.1 ms           16 bytes_per_second=7.0834G/s items_per_second=7.60575G/s
device,merge_sort,uint8_t,empty_type,default_config/manual_time                                          44.0 ms         44.0 ms           16 bytes_per_second=7.10657G/s items_per_second=7.63062G/s
device,merge_sort,rocprim::half,empty_type,default_config/manual_time                                    51.3 ms         51.3 ms           14 bytes_per_second=12.1757G/s items_per_second=6.53675G/s
device,merge_sort,short,empty_type,default_config/manual_time                                            53.7 ms         53.7 ms           13 bytes_per_second=11.6407G/s items_per_second=6.24954G/s
device,merge_sort,int,float,default_config/manual_time                                                    112 ms          112 ms            6 bytes_per_second=22.3192G/s items_per_second=2.99563G/s
device,merge_sort,int64_t,double,default_config/manual_time                                               191 ms          191 ms            4 bytes_per_second=26.1566G/s items_per_second=1.75534G/s
device,merge_sort,int8_t,int8_t,default_config/manual_time                                               62.6 ms         62.6 ms           11 bytes_per_second=9.98969G/s items_per_second=5.36318G/s
device,merge_sort,uint8_t,uint8_t,default_config/manual_time                                             62.3 ms         62.3 ms           11 bytes_per_second=10.0284G/s items_per_second=5.38395G/s
device,merge_sort,rocprim::half,rocprim::half,default_config/manual_time                                 94.3 ms         94.2 ms            7 bytes_per_second=13.2623G/s items_per_second=3.56008G/s
device,merge_sort,short,short,default_config/manual_time                                                 80.4 ms         80.4 ms            9 bytes_per_second=15.5507G/s items_per_second=4.17436G/s
device,merge_sort,int,custom_type<float,float>,default_config/manual_time                                 151 ms          151 ms            5 bytes_per_second=24.7727G/s items_per_second=2.21663G/s
device,merge_sort,int64_t,custom_type<double,double>,default_config/manual_time                           343 ms          343 ms            2 bytes_per_second=21.8573G/s items_per_second=977.878M/s
device,merge_sort,custom_type<double,double>,custom_type<double,double>,default_config/manual_time        515 ms          515 ms            1 bytes_per_second=19.4139G/s items_per_second=651.422M/s
device,merge_sort,custom_type<int,int>,custom_type<double,double>,default_config/manual_time              334 ms          334 ms            2 bytes_per_second=22.4305G/s items_per_second=1003.53M/s
device,merge_sort,custom_type<int,int>,custom_type<char,double>,default_config/manual_time                344 ms          344 ms            2 bytes_per_second=21.8064G/s items_per_second=975.6M/s
device,merge_sort,custom_type<int,int>,custom_type<int64_t,double>,default_config/manual_time             336 ms          336 ms            2 bytes_per_second=22.3156G/s items_per_second=998.384M/s

Process finished with exit code 0
```

After (this PR)
```cpp
device,merge_sort,int,empty_type,default_config/manual_time                                              56.9 ms         56.9 ms           12 bytes_per_second=21.956G/s items_per_second=5.89376G/s
device,merge_sort,int64_t,empty_type,default_config/manual_time                                           103 ms          103 ms            7 bytes_per_second=24.3041G/s items_per_second=3.26204G/s
device,merge_sort,int8_t,empty_type,default_config/manual_time                                           29.2 ms         29.2 ms           24 bytes_per_second=10.7065G/s items_per_second=11.496G/s
device,merge_sort,uint8_t,empty_type,default_config/manual_time                                          29.2 ms         29.2 ms           24 bytes_per_second=10.6916G/s items_per_second=11.4801G/s
device,merge_sort,rocprim::half,empty_type,default_config/manual_time                                    34.6 ms         34.6 ms           20 bytes_per_second=18.0611G/s items_per_second=9.69647G/s
device,merge_sort,short,empty_type,default_config/manual_time                                            40.7 ms         40.7 ms           17 bytes_per_second=15.366G/s items_per_second=8.24958G/s
device,merge_sort,int,float,default_config/manual_time                                                    102 ms          102 ms            7 bytes_per_second=24.5886G/s items_per_second=3.30022G/s
device,merge_sort,int64_t,double,default_config/manual_time                                               189 ms          189 ms            4 bytes_per_second=26.5083G/s items_per_second=1.77894G/s
device,merge_sort,int8_t,int8_t,default_config/manual_time                                               49.0 ms         49.0 ms           14 bytes_per_second=12.7582G/s items_per_second=6.84951G/s
device,merge_sort,uint8_t,uint8_t,default_config/manual_time                                             48.8 ms         48.8 ms           14 bytes_per_second=12.8172G/s items_per_second=6.88119G/s
device,merge_sort,rocprim::half,rocprim::half,default_config/manual_time                                 74.4 ms         74.4 ms            9 bytes_per_second=16.8051G/s items_per_second=4.5111G/s
device,merge_sort,short,short,default_config/manual_time                                                 66.4 ms         66.4 ms           11 bytes_per_second=18.8177G/s items_per_second=5.05132G/s
device,merge_sort,int,custom_type<float,float>,default_config/manual_time                                 149 ms          149 ms            5 bytes_per_second=25.1028G/s items_per_second=2.24616G/s
device,merge_sort,int64_t,custom_type<double,double>,default_config/manual_time                           292 ms          292 ms            2 bytes_per_second=25.7052G/s items_per_second=1.15003G/s
device,merge_sort,custom_type<double,double>,custom_type<double,double>,default_config/manual_time        496 ms          496 ms            2 bytes_per_second=20.1582G/s items_per_second=676.398M/s
device,merge_sort,custom_type<int,int>,custom_type<double,double>,default_config/manual_time              287 ms          287 ms            2 bytes_per_second=26.1547G/s items_per_second=1.17014G/s
device,merge_sort,custom_type<int,int>,custom_type<char,double>,default_config/manual_time                297 ms          296 ms            2 bytes_per_second=25.287G/s items_per_second=1.13132G/s
device,merge_sort,custom_type<int,int>,custom_type<int64_t,double>,default_config/manual_time             287 ms          287 ms            2 bytes_per_second=26.1023G/s items_per_second=1.1678G/s

Process finished with exit code 0
```